requirements.txt
loguru
pyotp
qrcode
tk
psycopg2-binary
twilio
